feat(cmi5_player): PR go back
move block of code from RC5Player that is dealing with navigation to a more appropriate hook (useSlideNavigation )
